Add hpssa module to proliantutils
This commit introduces the hpssa module. The
module has the ability to parse the current
raid configuration and hardware details of the
server. It also has a manager which has ability
to create/delete/get raid configuration.

For now, create is limited to machine-dependent
configuration.  Each logical disk to be created
will require the controller and physical disks to
be used.

Change-Id: Ib58620d8d4f47eb55c55c6abb5d3b294623205bf