HPSSA: Add functional tests for create
This commit adds some functional tests for create and
delete RAID config.  This needs to be run on Proliant
servers in a linux environment (IPA ramdisk).

NOTE: For now most of the stuffs are hardcoded and running
functional test will try deleting all the logical drives from
the selected controller.

Change-Id: Ia6b3e2f33094dd68bb6cc8a80998bcced5ab53d0