Add proper error message for extraction failure
This is a fix to provide a proper error message when extracting a
firmware file from compact format .rpm file fails. Commands like
`rpm2cpio` and `cpio` are required to extract the rpm files. It adds
a proper error message if those are not present as part of default
installation of linux.

Closes-Bug: #1569900
Change-Id: I401797afa59721bb3f8d0bbda56d11552cc83e9a