Add support to sort values of raid_config for 'share_physical_disks'
This commit adds the mechanism to sort the raid_config['logical_disks']
based on the number of physical disks needed to create the shared
logical volume.

Currenly it is sorted based on the 'size_gb' which doesn't allow
to share two logical volumes, if the second volume needs more disk
than the first volume.

Change-Id: I6117ae686d7530542aa3a338c2e57d5658b4f739