Fix on ris for very quick firmware update
When iteratively checking for the firmware flashing progress in case
of ris, there is a corner case wherein if the firmware file flashing
happens blazingly fast the logic might miss the finished status even
after exhausting its multiple progress check attempts. So, this fix
is provided to avoid such kind of scenarios.

Change-Id: If611ec81ad4361842f10cd2c2304652c14ea4733