Stop hosting coverage results in htmlcov dir
We now have a new Jenkins check job (code-coverage) running as
part of any proliantutils patch submitted upstream; and the dir
which is hosted as part of the coverage-job results is 'cover',
and not 'htmlcov' which we use to place the html output of the
code coverage results. As part of 'cover' dir we place the html
results in which the coverage takes into account the unit test
files also. Ideally code coverage should include only the
source files.

This patch addresses this concern by not using the htmlcov dir
any more and will use the cover dir only to host the actual
coverage results (i.e. devoid of unit test files).

Change-Id: I7cba8d4f0f2b9231b5f04b93b8a012889dd04c10