Initial redfish commit
This patch is again targeted towards support of consuming
Redfish APIs. This introduces the initial commit for redfish
related files. Till now, the redfish code base is **not** yet
instrumented to be called from ``ilo.client.IloClient`` object.
A separate patch will be submitted to instrument this redfish
code for Gen10. Till that this patch turns out to be a
pre-enablement of redfish on proliantutils.

Change-Id: Ib7f7596053543c3ebfa115ace3b56979ea5d6a36
Partial-Bug: 1646685