Instrumenting redfish into proliantutils' client
This patch is again targeted towards support of consuming
Redfish APIs. This introduces the instrumentation of redfish
object's method invocation from ``ilo.client.IloClient``
object. Till now, the client was instrumented to make the calls
in the following way::

  When we make a call like this:

    client_obj = client.IloClient(...)
	client_obj.method()

  Then it is translated as follows:

    if (model is Gen9) and (method is one of supported_ris_methods):
	  then ris_obj.method()
	else:
	  then ribcl_obj.method()

As part of this patch the translation logic gets changed to
this:

    if (model is Gen9) and (method is in supported_ris_methods):
	  then ris_obj.method()
	else if (model is Gen10) and (method is in supported_redfish_methods):
	  then redfish_obj.method()
	else:
	  then ribcl_obj.method()

Hence, in short this patch turns out to be the turnkey of
redfish support for Gen10 machines on proliantutils.

Change-Id: Iaf261a17da137a6fa49e62d9ff2a7cf34c97c2e3
Partial-Bug: 1646685