Remove sushy manager hack
Earlier we had included a temporary hack [0] for manager to pass
in 'manager_mock.py' file till the point a new Sushy version is
released with manager feature. This is to remove that temporary
hack as Sushy had been released recently.

Also pegged the version of Sushy to >=1.0.0 in requirements.

[0] https://review.openstack.org/#/c/473719/

Change-Id: I8eb2389730bb5d65338d2562fe2a9462bafc3479