Changes to support Sushy 2.x.x
The new Sushy library 2.x.x. will have changes in behaviour of the
'refresh()' method of ResourceBase. This commit takes care of changes
in prolintutils to keep prolintutils aligned to those changes in
Sushy.

Right now the Zuul jobs will fail as Sushy 2.x.x library with the new
changes as described is still not released and wont be available to
get installed on virtual environments. I have personally checked the
unit test cases running locally by linking with the new Sushy code

Change-Id: I1088d1b357bfe5c7cd76185bc4da484471ac3f6c