Fixes session issue for Gen10 servers
For some of the Gen 10 servers there is limit on the no.of session
which can be created on the ILO. If we are trying to create a new
session after the limit is reached on ILO, it returns code 401 and
this response "The Redfish System "1" was not found. Error HTTP GET
https://172.17.1.89/redfish/v1/Systems/1 returned code 401.".
When we use basic authentication at the time of sushy object creation,
new session are not created on ILO. So using basic authentication
when we are creating a sushy object which is used in redfish system
calls. This fixes the session issue for Gen10 servers.

Change-Id: Ic13da26a77863f8c383fa8ba185e05a8fc2e9fcf
Closes-Bug: #1764395