Revert "Add NIC inspection for Gen9"
This reverts commit 7286d7019c53ff55431690e07df02029acd171b2.
The fix is reverted because iLO4 behaviour is not consistent in
reporting the NIC data even for MCTP compliant NIC cards.
And iLO4 does not has the intelligence to detect the
NIC health status out-of-band for non-MCTP compliant NIC cards.
It reports the correct status only when AMS is running on the
server which requires OS to be present onto the system.

Change-Id: I02c27d820b27bd532f23ab7ef276b39f868f5205